added user validations
