+++ src/layouts/Main/components/Sidebar/index.jsx
@@ -4,42 +4,19 @@ import { Collapse, colors, Drawer, List, ListItem, makeStyles, SvgIcon, Typograp
 import {
     ArrowDropDown,
     ArrowDropUp,
-    ContactMail,
-    Dashboard,
-    SettingsApplications,
-    Category,
-    AccountTree,
-    ChatBubble,
-    Storage,
+    Category, 
 } from "@material-ui/icons";
-import CalendarTodayIcon from '@material-ui/icons/CalendarToday';
-import LibraryBooksIcon from '@material-ui/icons/LibraryBooks';
-import MenuBookIcon from '@material-ui/icons/MenuBook';
-
+import { adminList, siteOwnerList, memberList } from "./navLists"
 import cn from "classnames";
 import Scrollbar from "react-custom-scrollbars-2";
-
 import { computeHeight, getInitials } from "../../../../shared/utilities/common.util";
 
-const navList = [
-    { id: "home", path: "/home", label: 'Home', icon: Dashboard },
-    { id: "communications", path: "/communications", label: 'Communication', icon: ChatBubble },
-    {
-        id: "project_setting", label: "Project Setting", icon: SettingsApplications, subMenu: [
-            { id: "groups", path: "/groups", label: 'Groups' },
-            { id: "roles", path: "/roles", label: 'Role' },
-            { id: "permissions", path: "/permissions", label: 'Permission' },
-        ]
-    },
-    { id: "projectInformation", path: "/projectInformation", label: 'Project Information', icon: AccountTree },
-    { id: "contacts_id", path: "/contacts", label: 'Members', icon: ContactMail },
-    { id: "calander", path: "/calander", label: 'Calander', icon: CalendarTodayIcon },
-    { id: "documentManager", path: "/storages", label: 'File Manager', icon: Storage },
-    { id: "library", path: "/library", label: 'Library', icon: MenuBookIcon },
-    { id: "report", path: "/report", label: 'Report', icon: LibraryBooksIcon },
-    
-];
+var navList = memberList
 
+const user = localStorage.getItem("user");
+ if ( user === "site_owner" ){ navList = siteOwnerList }
+ else if ( user === "project_admin" ){ navList = adminList }
+        
 const useStyles = makeStyles((theme) => ({
     flexGrow: { flexGrow: 1 },
     drawer: {
@@ -157,7 +134,7 @@ const Sidebar = ({ openSidebar }) => {
             [menuId]: !collapseState[menuId]
         });
     };
-
+            
     const getSidebarNav = (navList, isSubList = false, parentIndex = '') => {
         return (
             <List component="ul" disablePadding className={cn({ [classes.navList]: !isSubList })}>
