+++ src/pages/Home/index.jsx
@@ -1,51 +1,21 @@
 import Grid from "@material-ui/core/Grid";
 import { Container } from "components";
-import AvailableProjects from "./Components/AvailableProjects";
-import Calander from "./Components/Calander";
-import Communication from "./Components/Communication";
-import DocumentManager from "./Components/DocumentManager";
-import Library from "./Components/Library";
-import ProjectInformation from "./Components/ProjectInformation";
-import ProjectSettings from "./Components/ProjectSettings";
-import Reports from "./Components/Reports";
 import SelectProject from "./Components/SelectProject";
-import ChatModule from './Components/ChatModule'
-import ChatBox from './Components/ChatBox'
-
+import ProjectAdminView from "./ProjectAdminView";
+import SiteOwnerView from "./SiteOwnerView";
+import MemberView from "./MemberView";
 const Home = ({ history, location }) => {
-    
+    const user = localStorage.getItem("user");
     const project_id = localStorage.getItem('project_id');
     if (project_id) {
         return (
             <Container title="Home">
-                <Grid container spacing={2}>
-                    <Grid item lg={3} sm={6} xl={3} xs={12}>
-                        <AvailableProjects history={history} location={location}></AvailableProjects>
-                    </Grid>
-                    <Grid item lg={3} sm={6} xl={3} xs={12}>
-                        <ProjectInformation history={history} location={location}/>
-                    </Grid>
-                    <Grid item lg={3} sm={6} xl={3} xs={12}>
-                        <ProjectSettings history={history} location={location} />
-                    </Grid>
-                    <Grid item lg={3} sm={6} xl={3} xs={12}>
-                        <Communication history={history} location={location} />
-                    </Grid>
-                    <Grid item lg={3} sm={6} xl={3} xs={12}>
-                        <Calander history={history} location={location}/>
-                    </Grid>
-                    <Grid item lg={3} sm={6} xl={3} xs={12}>
-                        <DocumentManager history={history} location={location}/>
-                    </Grid>
-                    <Grid item lg={3} sm={6} xl={3} xs={12}>
-                        <Library history={history} location={location}/>
-                    </Grid>
-                    <Grid item lg={3} sm={6} xl={3} xs={12}>
-                        <Reports history={history} location={location}/>
-                    </Grid>
-                </Grid>
-               <ChatModule></ChatModule>
-               {/* <ChatBox></ChatBox> */}
+            {user==="site_owner" &&   //site-owner and admin view are currently same till we add site-owner controls
+            <SiteOwnerView history={history}></SiteOwnerView>}      
+            {user==="project_admin" &&              
+             <ProjectAdminView history={history}></ProjectAdminView>}
+             {user==="project_member" && 
+             <MemberView history={history}></MemberView>}
             </Container>
         );
     }
